fixing unit tests after refactoring
