+++ packages/destination-actions/src/destinations/engage/utils/index.ts
@@ -1,2 +1,8 @@
 export * from './EngageActionPerformer'
 export * from './operationTracking'
+export * from './MessageLogger'
+export * from './MessageStats'
+export * from './track'
+export * from './isDestinationActionService'
+export * from './TwilioApiError'
+export * from './IntegrationErrorWrapper'
