+++ packages/destination-actions/src/destinations/engage/utils/track.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ContextFromDecorator, OperationDecorator } from '../../utils/operationTracking'
+import { ContextFromDecorator, OperationDecorator } from './operationTracking'
 import { MessageStats } from './MessageStats'
 import { MessageLogger } from './MessageLogger'
 import { IntegrationErrorWrapper } from './IntegrationErrorWrapper'
